Function definition: use uint16_t for nTriggers.

+++ core/api/api.h
@@ -291,7 +291,7 @@ namespace pxar {
      *  over "nTriggers" triggers
      */
     std::vector< std::pair<uint8_t, std::vector<pixel> > > getPulseheightVsDAC(std::string dacName, uint8_t dacMin, uint8_t dacMax, 
-									       uint16_t flags = 0, uint32_t nTriggers = 16);
+									       uint16_t flags = 0, uint16_t nTriggers = 16);
 
     /** Method to scan a DAC range and measure the efficiency
      *
@@ -300,7 +300,7 @@ namespace pxar {
      *  pixel. Efficiency == 1 for nhits == nTriggers
      */
     std::vector< std::pair<uint8_t, std::vector<pixel> > > getEfficiencyVsDAC(std::string dacName, uint8_t dacMin, uint8_t dacMax, 
-					  uint16_t flags = 0, uint32_t nTriggers=16);
+					  uint16_t flags = 0, uint16_t nTriggers=16);
 
     /** Method to scan a DAC range and measure the pixel threshold
      *
@@ -310,7 +310,7 @@ namespace pxar {
      *
      *  The threshold is calculated as the 0.5 value of the s-curve of the pixel.
      */
-    std::vector< std::pair<uint8_t, std::vector<pixel> > > getThresholdVsDAC(std::string dacName, std::string dac2name, uint8_t dac2min, uint8_t dac2max, uint16_t flags = 0, uint32_t nTriggers=16);
+    std::vector< std::pair<uint8_t, std::vector<pixel> > > getThresholdVsDAC(std::string dacName, std::string dac2name, uint8_t dac2min, uint8_t dac2max, uint16_t flags = 0, uint16_t nTriggers=16);
 
     /** Method to scan a DAC range and measure the pixel threshold
      *
@@ -323,7 +323,7 @@ namespace pxar {
      *
      *  The threshold is calculated as the 0.5 value of the s-curve of the pixel.
      */
-    std::vector< std::pair<uint8_t, std::vector<pixel> > > getThresholdVsDAC(std::string dac1name, uint8_t dac1min, uint8_t dac1max, std::string dac2name, uint8_t dac2min, uint8_t dac2max, uint16_t flags, uint32_t nTriggers);
+    std::vector< std::pair<uint8_t, std::vector<pixel> > > getThresholdVsDAC(std::string dac1name, uint8_t dac1min, uint8_t dac1max, std::string dac2name, uint8_t dac2min, uint8_t dac2max, uint16_t flags, uint16_t nTriggers);
 
     /** Method to scan a 2D DAC-Range (DAC1 vs. DAC2) and measure the
      *  pulse height
@@ -334,7 +334,7 @@ namespace pxar {
      */
     std::vector< std::pair<uint8_t, std::pair<uint8_t, std::vector<pixel> > > > getPulseheightVsDACDAC(std::string dac1name, uint8_t dac1min, uint8_t dac1max, 
 					      std::string dac2name, uint8_t dac2min, uint8_t dac2max, 
-					      uint16_t flags = 0, uint32_t nTriggers=16);
+					      uint16_t flags = 0, uint16_t nTriggers=16);
 
     /** Method to scan a 2D DAC-Range (DAC1 vs. DAC2) and measure the efficiency
      *
@@ -344,21 +344,21 @@ namespace pxar {
      */
     std::vector< std::pair<uint8_t, std::pair<uint8_t, std::vector<pixel> > > > getEfficiencyVsDACDAC(std::string dac1name, uint8_t dac1min, uint8_t dac1max, 
 					     std::string dac2name, uint8_t dac2min, uint8_t dac2max, 
-					     uint16_t flags = 0, uint32_t nTriggers=16);
+					     uint16_t flags = 0, uint16_t nTriggers=16);
 
     /** Method to get a map of the pulse height
      *
      *  Returns a vector of pixels, with the value of the pixel struct being
      *  the averaged pulse height over "nTriggers" triggers
      */
-    std::vector<pixel> getPulseheightMap(uint16_t flags = 0, uint32_t nTriggers=16);
+    std::vector<pixel> getPulseheightMap(uint16_t flags = 0, uint16_t nTriggers=16);
 
     /** Method to get a map of the efficiency
      *
      *  Returns a vector of pixels, with the value of the pixel struct being
      *  the number of hits in that pixel. Efficiency == 1 for nhits == nTriggers
      */
-    std::vector<pixel> getEfficiencyMap(uint16_t flags = 0, uint32_t nTriggers=16);
+    std::vector<pixel> getEfficiencyMap(uint16_t flags = 0, uint16_t nTriggers=16);
 
     /** Method to get a map of the pixel threshold
      *
@@ -370,7 +370,7 @@ namespace pxar {
      *
      *  The threshold is calculated as the 0.5 value of the s-curve of the pixel.
      */
-    std::vector<pixel> getThresholdMap(std::string dacName, uint8_t dacMin, uint8_t dacMax, uint16_t flags, uint32_t nTriggers);
+    std::vector<pixel> getThresholdMap(std::string dacName, uint8_t dacMin, uint8_t dacMax, uint16_t flags, uint16_t nTriggers);
 
     /** Method to get a map of the pixel threshold
      *
@@ -379,7 +379,7 @@ namespace pxar {
      *
      *  The threshold is calculated as the 0.5 value of the s-curve of the pixel.
      */
-    std::vector<pixel> getThresholdMap(std::string dacName, uint16_t flags = 0, uint32_t nTriggers=16);
+    std::vector<pixel> getThresholdMap(std::string dacName, uint16_t flags = 0, uint16_t nTriggers=16);
 
     // FIXME missing documentation
     int32_t getReadbackValue(std::string parameterName);
